split spec files
